[F] warnings

+++ QAKit/QAKit/NSFoundation/NSArray+kit.m
@@ -27,8 +27,8 @@
     NSUInteger count = [workingArray count];
     for (NSUInteger i = 0; i < count; ++i) {
         // Select a random element between i and end of array to swap with.
-        int nElements = count - i;
-        int n = (arc4random() % nElements) + i;
+        NSUInteger nElements = count - i;
+        NSUInteger n = (arc4random() % nElements) + i;
         [workingArray exchangeObjectAtIndex:i withObjectAtIndex:n];
     }
